NEWRELIC-4420 removed the old async-hooks tests, and renamed the new-promise-tracking async-hooks tests to async-hooks as this is now the default. also pared down the native-promises tests as they were duplicates of the new async-hooks.js tests
