+++ test/integration/core/native-promises/async-hooks.js
@@ -4,26 +4,31 @@
  */
 
 'use strict'
+
 const test = require('tap').test
 const helper = require('../../../lib/agent_helper')
 const asyncHooks = require('async_hooks')
 
-function testSegments(t, segmentMap) {
-  global.gc()
-  // Give the gc some time to work.
-  setTimeout(function () {
-    t.notOk(segmentMap.size, 'segments should be cleared after gc')
-    t.end()
-  }, 10)
-}
-
 test('await', function (t) {
   const { agent } = setupAgent(t)
+
   helper.runInTransaction(agent, async function (txn) {
     let transaction = agent.getTransaction()
     t.equal(transaction && transaction.id, txn.id, 'should start in a transaction')
 
-    await Promise.resolve("i'll be back")
+    const segmentMap = require('../../../../lib/instrumentation/core/async_hooks').segmentMap
+
+    const promise = new Promise((resolve) => {
+      // don't immediately resolve so logic can kick in.
+      setImmediate(resolve)
+    })
+
+    // There may be extra promises in play
+    const promiseId = [...segmentMap.keys()].pop()
+
+    await promise
+
+    t.notOk(segmentMap.has(promiseId), 'should have removed segment for promise after resolve')
 
     transaction = agent.getTransaction()
     t.equal(
@@ -32,17 +37,13 @@ test('await', function (t) {
       'should resume in the same transaction after await'
     )
 
-    const segmentMap = require('../../../../lib/instrumentation/core/async_hooks').segmentMap
     txn.end()
-    // Segments won't be cleared till a gc cycle clears the promises
-    // they are related with.
-    t.ok(segmentMap.size, 'segments should still be in the map')
-    if (global.gc) {
-      // Unroll all the stack frames to let go of the refs to the
-      // promises we want to gc, then call the segment tester.
-      return setImmediate(testSegments, t, segmentMap)
-    }
-    t.end()
+
+    // Let the loop iterate to clear the microtask queue
+    setImmediate(() => {
+      t.equal(segmentMap.size, 0, 'should clear segments after all promises resolved')
+      t.end()
+    })
   })
 })
 
@@ -70,9 +71,9 @@ test("the agent's async hook", function (t) {
     const { agent } = setupAgent(t)
     helper.runInTransaction(agent, function () {
       t.doesNotThrow(function () {
-        new Promise(function (res) {
-          res()
-          res()
+        new Promise(function (resolve) {
+          resolve()
+          resolve()
         }).then(t.end)
       })
     })
@@ -81,13 +82,13 @@ test("the agent's async hook", function (t) {
   t.test('does not restore a segment for a resource created outside a transaction', function (t) {
     const { agent, contextManager } = setupAgent(t)
 
-    const res = new TestResource(1)
+    const testResource = new TestResource(1)
     helper.runInTransaction(agent, function () {
       const root = contextManager.getContext()
       const segmentMap = require('../../../../lib/instrumentation/core/async_hooks').segmentMap
 
       t.equal(segmentMap.size, 0, 'no segments should be tracked')
-      res.doStuff(function () {
+      testResource.doStuff(function () {
         t.ok(contextManager.getContext(), 'should be in a transaction')
         t.equal(
           contextManager.getContext().name,
@@ -101,11 +102,12 @@ test("the agent's async hook", function (t) {
 
   t.test('restores context in inactive transactions', function (t) {
     const { agent, contextManager } = setupAgent(t)
+
     helper.runInTransaction(agent, function (txn) {
-      const res = new TestResource(1)
+      const testResource = new TestResource(1)
       const root = contextManager.getContext()
       txn.end()
-      res.doStuff(function () {
+      testResource.doStuff(function () {
         t.equal(
           contextManager.getContext(),
           root,
@@ -116,7 +118,55 @@ test("the agent's async hook", function (t) {
     })
   })
 
-  t.test('parent promises persist perspective to problematic progeny', function (t) {
+  /**
+   * Represents same test as 'parent promises persist perspective to problematic progeny'
+   * from async_hooks.js.
+   *
+   * This specific use case is not currently supported with the implementation that clears
+   * segment references on promise resolve.
+   */
+  t.test(
+    'parent promises that are already resolved DO NOT persist to continuations ' +
+      'scheduled after a timer async hop.',
+    function (t) {
+      const { agent } = setupAgent(t)
+      const tasks = []
+      const intervalId = setInterval(() => {
+        while (tasks.length) {
+          tasks.pop()()
+        }
+      }, 10)
+
+      t.teardown(() => {
+        clearInterval(intervalId)
+      })
+
+      helper.runInTransaction(agent, function (txn) {
+        t.ok(txn, 'transaction should not be null')
+
+        const p = Promise.resolve()
+
+        tasks.push(() => {
+          p.then(() => {
+            const tx = agent.getTransaction()
+            t.not(
+              tx ? tx.id : null,
+              txn.id,
+              'If this failed, this use case now works! Time to switch to "t.equal"'
+            )
+            t.end()
+          })
+        })
+      })
+    }
+  )
+
+  /**
+   * Variation of 'parent promises persist perspective to problematic progeny' from async_hooks.js.
+   *
+   * For unresolved parent promises, persistance should stil work as expected.
+   */
+  t.test('unresolved parent promises persist perspective to problematic progeny', function (t) {
     const { agent } = setupAgent(t)
     const tasks = []
     const intervalId = setInterval(() => {
@@ -132,19 +182,32 @@ test("the agent's async hook", function (t) {
     helper.runInTransaction(agent, function (txn) {
       t.ok(txn, 'transaction should not be null')
 
-      const p = Promise.resolve()
+      let parentResolve = null
+      const p = new Promise((resolve) => {
+        parentResolve = resolve
+      })
 
       tasks.push(() => {
         p.then(() => {
           const tx = agent.getTransaction()
           t.equal(tx ? tx.id : null, txn.id)
+
           t.end()
         })
+
+        // resolve parent after continuation scheduled
+        parentResolve()
       })
     })
   })
 
-  t.test('maintains transaction context', function (t) {
+  /**
+   * Represents same test as 'maintains transaction context' from async_hooks.js.
+   *
+   * Combination of a timer that does not propagate state and the new resolve
+   * mechanism that clears (and sets hook as active) causes this to fail.
+   */
+  t.test('DOES NOT maintain transaction context', function (t) {
     const { agent } = setupAgent(t)
     const tasks = []
     const intervalId = setInterval(() => {
@@ -165,6 +228,7 @@ test("the agent's async hook", function (t) {
       const wrapperTwo = agent.tracer.bindFunction(function () {
         return two()
       }, segment)
+
       const wrapperThree = agent.tracer.bindFunction(function () {
         return three()
       }, segment)
@@ -181,7 +245,11 @@ test("the agent's async hook", function (t) {
         tasks.push(() => {
           next().then(() => {
             const tx = agent.getTransaction()
-            t.equal(tx ? tx.id : null, txn.id)
+            t.not(
+              tx ? tx.id : null,
+              txn.id,
+              'If this failed, this use case now works! Time to switch to "t.equal"'
+            )
             resolve()
           })
         })
@@ -197,6 +265,76 @@ test("the agent's async hook", function (t) {
 
       function nextTwo() {
         return Promise.resolve(wrapperThree())
+      }
+
+      function three() {}
+    })
+  })
+
+  t.test('maintains transaction context for unresolved promises', function (t) {
+    const { agent } = setupAgent(t)
+    const tasks = []
+    const intervalId = setInterval(() => {
+      while (tasks.length) {
+        tasks.pop()()
+      }
+    }, 10)
+
+    t.teardown(() => {
+      clearInterval(intervalId)
+    })
+
+    helper.runInTransaction(agent, function (txn) {
+      t.ok(txn, 'transaction should not be null')
+      const segment = txn.trace.root
+      agent.tracer.bindFunction(one, segment)()
+
+      const wrapperTwo = agent.tracer.bindFunction(function () {
+        return two()
+      }, segment)
+
+      const wrapperThree = agent.tracer.bindFunction(function () {
+        return three()
+      }, segment)
+
+      function one() {
+        return new Promise(executor).then(() => {
+          const tx = agent.getTransaction()
+          t.equal(tx ? tx.id : null, txn.id)
+          t.end()
+        })
+      }
+
+      function executor(resolve) {
+        setImmediate(() => {
+          next().then(() => {
+            const tx = agent.getTransaction()
+            t.equal(tx ? tx.id : null, txn.id)
+            resolve()
+          })
+        })
+      }
+
+      function next() {
+        return new Promise((resolve) => {
+          const val = wrapperTwo()
+          setImmediate(() => {
+            resolve(val)
+          })
+        })
+      }
+
+      function two() {
+        return nextTwo()
+      }
+
+      function nextTwo() {
+        return new Promise((resolve) => {
+          const val = wrapperThree()
+          setImmediate(() => {
+            resolve(val)
+          })
+        })
       }
 
       function three() {}
@@ -282,6 +420,7 @@ test("the agent's async hook", function (t) {
 
   t.test('handles multientry callbacks correctly', function (t) {
     const { agent, contextManager } = setupAgent(t)
+
     const segmentMap = require('../../../../lib/instrumentation/core/async_hooks').segmentMap
     helper.runInTransaction(agent, function () {
       const root = contextManager.getContext()
@@ -320,13 +459,11 @@ test("the agent's async hook", function (t) {
           'runInAsyncScope should restore the segment active when a callback was called'
         )
       })
-
       t.equal(
         contextManager.getContext().name,
         root.name,
         'root should be restored after we are finished'
       )
-
       resA.doStuff(() => {
         t.equal(
           contextManager.getContext().name,
@@ -336,6 +473,68 @@ test("the agent's async hook", function (t) {
       })
     })
   })
+
+  t.test(
+    'cleans up unresolved promises on destroy',
+    { skip: process.env.NEW_RELIC_FEATURE_FLAG_UNRESOLVED_PROMISE_CLEANUP === 'false' },
+    (t) => {
+      const { agent } = setupAgent(t)
+      const segmentMap = require('../../../../lib/instrumentation/core/async_hooks').segmentMap
+
+      helper.runInTransaction(agent, () => {
+        /* eslint-disable no-unused-vars */
+        let promise = unresolvedPromiseFunc()
+
+        t.equal(segmentMap.size, 1, 'segment map should have 1 element')
+
+        promise = null
+
+        global.gc && global.gc()
+
+        setImmediate(() => {
+          t.equal(segmentMap.size, 0, 'segment map should clean up unresolved promises on destroy')
+          t.end()
+        })
+      })
+
+      function unresolvedPromiseFunc() {
+        return new Promise(() => {})
+      }
+    }
+  )
+
+  t.test(
+    'does not clean up unresolved promises on destroy when `unresolved_promise_cleanup` is set to false',
+    { skip: process.env.NEW_RELIC_FEATURE_FLAG_UNRESOLVED_PROMISE_CLEANUP !== 'false' },
+    (t) => {
+      const { agent } = setupAgent(t)
+      const segmentMap = require('../../../../lib/instrumentation/core/async_hooks').segmentMap
+
+      helper.runInTransaction(agent, () => {
+        /* eslint-disable no-unused-vars */
+        let promise = unresolvedPromiseFunc()
+
+        t.equal(segmentMap.size, 1, 'segment map should have 1 element')
+
+        promise = null
+
+        global.gc && global.gc()
+
+        setImmediate(() => {
+          t.equal(
+            segmentMap.size,
+            1,
+            'segment map should not clean up unresolved promise on destroy'
+          )
+          t.end()
+        })
+      })
+
+      function unresolvedPromiseFunc() {
+        return new Promise(() => {})
+      }
+    }
+  )
 })
 
 function checkCallMetrics(t, testMetrics) {
@@ -400,8 +599,8 @@ test('promise hooks', function (t) {
   hook.enable()
 
   t.test('are only called once during the lifetime of a promise', function (t) {
-    new Promise(function (res) {
-      setTimeout(res, 10)
+    new Promise(function (resolve) {
+      setTimeout(resolve, 10)
     }).then(function () {
       setImmediate(checkCallMetrics, t, testMetrics)
     })
@@ -421,5 +620,8 @@ function setupAgent(t) {
     helper.unloadAgent(agent)
   })
 
-  return { agent, contextManager }
+  return {
+    agent,
+    contextManager
+  }
 }
