+++ test/integration/core/native-promises/native-promises.js
@@ -11,196 +11,6 @@ const semver = require('semver')
 const helper = require('../../../lib/agent_helper')
 const asyncHooks = require('async_hooks')
 
-const skipAsyncHooks = process.env.NEW_RELIC_FEATURE_FLAG_ASYNC_LOCAL_CONTEXT
-test('Async-hooks + timer instrumentation based tracking', { skip: skipAsyncHooks }, (t) => {
-  t.autoend()
-
-  const config = {
-    feature_flag: {
-      await_support: true,
-      async_local_context: false
-    }
-  }
-
-  createPromiseTests(t, config)
-
-  t.test('parent promises persist perspective to problematic progeny', (t) => {
-    const { agent } = setupAgent(t, config)
-    const tasks = []
-    const intervalId = setInterval(() => {
-      while (tasks.length) {
-        tasks.pop()()
-      }
-    }, 10)
-
-    t.teardown(() => {
-      clearInterval(intervalId)
-    })
-
-    helper.runInTransaction(agent, function (txn) {
-      t.ok(txn, 'transaction should not be null')
-
-      const p = Promise.resolve()
-
-      tasks.push(() => {
-        p.then(() => {
-          const tx = agent.getTransaction()
-          t.equal(tx ? tx.id : null, txn.id)
-          t.end()
-        })
-      })
-    })
-  })
-
-  t.test('maintains transaction context over contextless timer', (t) => {
-    const { agent } = setupAgent(t, config)
-    const tasks = []
-    const intervalId = setInterval(() => {
-      while (tasks.length) {
-        tasks.pop()()
-      }
-    }, 10)
-
-    t.teardown(() => {
-      clearInterval(intervalId)
-    })
-
-    helper.runInTransaction(agent, function (txn) {
-      t.ok(txn, 'transaction should not be null')
-
-      const segment = txn.trace.root
-      agent.tracer.bindFunction(one, segment)()
-
-      const wrapperTwo = agent.tracer.bindFunction(function () {
-        return two()
-      }, segment)
-      const wrapperThree = agent.tracer.bindFunction(function () {
-        return three()
-      }, segment)
-
-      function one() {
-        return new Promise(executor).then(() => {
-          const tx = agent.getTransaction()
-          t.equal(tx ? tx.id : null, txn.id)
-          t.end()
-        })
-      }
-
-      function executor(resolve) {
-        tasks.push(() => {
-          next().then(() => {
-            const tx = agent.getTransaction()
-            t.equal(tx ? tx.id : null, txn.id)
-            resolve()
-          })
-        })
-      }
-
-      function next() {
-        return Promise.resolve(wrapperTwo())
-      }
-
-      function two() {
-        return nextTwo()
-      }
-
-      function nextTwo() {
-        return Promise.resolve(wrapperThree())
-      }
-
-      function three() {}
-    })
-  })
-
-  t.test('unresolved parent promises persist to problematic progeny', (t) => {
-    const { agent } = setupAgent(t, config)
-    const tasks = []
-    const intervalId = setInterval(() => {
-      while (tasks.length) {
-        tasks.pop()()
-      }
-    }, 10)
-
-    t.teardown(() => {
-      clearInterval(intervalId)
-    })
-
-    helper.runInTransaction(agent, function (txn) {
-      t.ok(txn, 'transaction should not be null')
-
-      let parentResolve = null
-      const p = new Promise((resolve) => {
-        parentResolve = resolve
-      })
-
-      tasks.push(() => {
-        p.then(() => {
-          const tx = agent.getTransaction()
-          t.equal(tx ? tx.id : null, txn.id)
-
-          t.end()
-        })
-
-        // resolve parent after continuation scheduled
-        parentResolve()
-      })
-    })
-  })
-
-  t.test('loses transaction context when no promise returned', function (t) {
-    const { agent } = setupAgent(t, config)
-    const tasks = []
-    const intervalId = setInterval(() => {
-      while (tasks.length) {
-        tasks.pop()()
-      }
-    }, 10)
-
-    t.teardown(() => {
-      clearInterval(intervalId)
-    })
-
-    helper.runInTransaction(agent, function (txn) {
-      t.ok(txn, 'transaction should not be null')
-      const segment = txn.trace.root
-      agent.tracer.bindFunction(one, segment)()
-
-      const wrapperTwo = agent.tracer.bindFunction(function () {
-        return two()
-      }, segment)
-
-      function one() {
-        return new Promise(executor).then(() => {
-          const tx = agent.getTransaction()
-          t.equal(tx ? tx.id : null, txn.id)
-          t.end()
-        })
-      }
-
-      function executor(resolve) {
-        tasks.push(() => {
-          next().then(() => {
-            const tx = agent.getTransaction()
-            // We know tx will be null here because no promise was returned
-            // If this test fails, that's actually a good thing,
-            // so throw a party/update Koa.
-            t.equal(tx, null)
-            resolve()
-          })
-        })
-      }
-
-      function next() {
-        return Promise.resolve(wrapperTwo())
-      }
-
-      function two() {
-        // No promise is returned to reinstate transaction context
-      }
-    })
-  })
-})
-
 const skipAsyncLocal = semver.satisfies(process.version, '<16.4.0')
 test('AsyncLocalStorage based tracking', { skip: skipAsyncLocal }, (t) => {
   t.autoend()
